requirements.txt
Django==1.8.1
djangorestframework==3.1.2
face-client==1.2.dev0
jsonfield==1.0.3
Pillow==2.8.1
